requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
sqlparse==0.4.4
typing_extensions==4.7.1
